Add files via upload
A customer in the Personalised Gifts Store is awarded a prize when their bill number is a 3-digit number and all the 3 digits are in sorted order. 
Examples: 379, 256, 973, 652, 225, 522 .... Help Nickson in identifying the prize winners.

Input Format:
Input consists of a number that corresponds to the bill number.

Output Format:
The output consists of a string that is either 'yes' or 'no'. The output is yes when the customer receives the prize and is no otherwise.

Sample Input 1:
565
Sample Output 1:
no

Sample Input 2:
620
Sample Output 2:
yes

Sample Input 3:
66
Sample Output 3:
no